backend/requirements.txt
absl-py==2.1.0
affine==2.4.0
alembic==1.16.2
aniso8601==10.0.1
astunparse==1.6.3
attrs==24.3.0
audioread==3.0.1
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-plugins==1.1.1
cligj==0.7.2
cloudinary==1.44.1
colorama==0.4.6
decorator==5.2.1
distlib==0.3.9
filelock==3.16.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
gast==0.6.0
gevent==25.5.1
gevent-websocket==0.10.1
google-pasta==0.2.0
greenlet==3.2.3
grpcio==1.66.2
h11==0.16.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
ml-dtypes==0.4.1
msgpack==1.1.1
namex==0.0.8
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
opt_einsum==3.4.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
patsy==1.0.1
pillow==11.2.1
platformdirs==4.3.6
pooch==1.8.2
protobuf==4.25.5
pycparser==2.22
Pygments==2.18.0
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.1
rasterio==1.4.3
regex==2024.9.11
reportlab==4.4.2
requests==2.32.3
scikit-learn==1.7.0
scipy==1.15.1
setuptools==75.1.0
simple-websocket==1.1.0
six==1.16.0
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.38
sqlparse==0.5.3
statsmodels==0.14.4
stripe==12.2.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.6.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
virtualenv==20.28.0
Werkzeug==3.1.3
wheel==0.44.0
wrapt==1.16.0
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
